app/templates_data/downloader_003/pyproject.toml
[project]
name = "nazurin"
version = "2.8.2"
description = "Images collection done right. A Telegram bot that helps you collect ACG illustrations."
readme = "README.md"
authors = [{ name = "y-young", email = "i@gpx.moe" }]
license-files = ["LICENSE"]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "aiofiles~=24.1.0",
    "aiogram~=3.26",
    "aiohttp~=3.13",
    "aiohttp-cors~=0.7.0",
    "aiohttp-socks~=0.10.1",
    "aiojobs==1.3.0",
    "async-lru~=2.0.2",
    "beautifulsoup4~=4.13.3",
    "cloudant~=2.15.0",
    "curl-cffi~=0.15.0",
    "dnspython~=2.6.1",
    "environs~=14.1.0",
    "firebase-admin~=7.4",
    "humanize~=4.11.0",
    "loguru~=0.7.3",
    "mega-py",
    "minio~=7.2.15",
    "pillow~=12.1.1",
    "pixivpy3~=3.7.2",
    "pybooru~=4.2.2",
    "pydantic~=2.10.6",
    "pydrive2~=1.21.3",
    "pymongo~=4.16",
    "tenacity~=9.0.0",
    "tinydb~=4.8.2",
    "typing-extensions>=4.12.2",
    "xclienttransaction==1.0.2",
]

[project.scripts]
nazurin = "nazurin:__main__"

[project.urls]
Homepage = "https://github.com/y-young/nazurin"
Documentation = "https://nazurin.readthedocs.io/"
Repository = "https://github.com/y-young/nazurin.git"
Issues = "https://github.com/y-young/nazurin/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
extend-select = [
    "ASYNC",  # flake8-async
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "COM",    # flake8-commas
    "DTZ",    # flake8-datetimez
    "E",      # pycodestyle
    "F",      # pyflakes
    "FBT",    # flake8-boolean-trap
    "FURB",   # refurb
    "G",      # flake8-logging-format
    "I",      # isort
    "INP",    # flake8-no-pep420
    "ISC",    # flake8-implicit-str-concat
    "N",      # pep8-naming
    "PERF",   # perflint
    "PIE",    # flake8-pie
    "PL",     # pylint
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "RSE102", # unnecessary-paren-on-raise-exception
    "RUF",    # ruff
    "SIM",    # flake8-simplify
    "T10",    # flake8-debugger
    "T20",    # flake8-print
    "TC",     # flake8-type-checking
    "UP",     # pyupgrade
    "W",      # pycodestyle,
    "YTT",    # flake8-2020
]
ignore = ["PERF203", "COM812", "RET504"]

[tool.uv]
constraint-dependencies = ["grpcio>=1.71.0"]

[tool.uv.sources]
mega-py = { git = "https://github.com/y-young/mega.py", rev = "6015359" }

[dependency-groups]
dev = ["pre-commit~=4.1.0", "pytest~=8.3.3", "rich>=14.0.0", "ruff~=0.15.6"]
